:nail_care: style(none): update @link url; organize types
