+++ sources/@roots/bud-build/src/index.ts
@@ -2,20 +2,20 @@
 // Licensed under the MIT license.
 
 /**
- * Compiler configuration builder
+ * Configuration builder
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
  */
 
 export {Build} from './Build'
+
 export {Loader} from './Loader'
 export {Item} from './Item'
 export {Rule} from './Rule'
 
-import * as items from './Build/items'
-import * as loaders from './Build/loaders'
-import * as rules from './Build/rules'
-export {items, loaders, rules}
+export * as items from './items'
+export * as loaders from './loaders'
+export * as rules from './rules'
