+++ sources/@roots/bud-cache/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * The {@link @roots/bud-cache#} package implements the {@link @roots/bud-framework#Cache | Cache interface}
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
