+++ sources/@roots/bud-compress/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Static asset compression extension
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
