+++ sources/@roots/bud-criticalcss/src/index.ts
@@ -8,7 +8,7 @@
  * @beta
  * This extension is under active development. But it should not be considered stable and there may be breaking changes.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
