+++ sources/@roots/bud-dashboard/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * The Dashboard package implements the Bud.Dashboard interface.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  *  @packageDocumentation
