+++ sources/@roots/bud-emotion/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * The `@roots/bud-emotion` package adds emotioncss support to {@link @roots/bud-framework# | the Bud}
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
