+++ sources/@roots/bud-esbuild/src/features/esbuild-js.ts
@@ -3,7 +3,7 @@ import type {Extension} from '@roots/bud-framework'
 export const jsFeature: Extension.Plugin = {
   name: '@roots/bud-esbuild/js',
 
-  boot: app => {
+  boot: async app => {
     app.build
       .setItem('esbuild-js', {
         loader: `esbuild`,
