+++ sources/@roots/bud-eslint/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Provides eslint integration for Bud.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
