+++ sources/@roots/bud-extensions/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * The {@link @roots/bud-extensions#} package implements the {@link @roots/bud-framework#Extensions | Extensions interface}
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  *  @packageDocumentation
