+++ sources/@roots/bud-framework/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * ⚡️ `@roots/bud-framework` - Extensible build tooling for modern web development
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
@@ -16,7 +16,6 @@ export * as Cache from './services/cache'
 export * as Compiler from './services/compiler'
 export * as Dashboard from './services/dashboard'
 export * as Env from './services/env'
-export * as Extension from './services/extensions'
 export * as Extensions from './services/extensions'
 export * as Hooks from './services/hooks'
 export * as Peers from './services/peers'
@@ -30,6 +29,8 @@ export {Bud} from './bud'
 export {ContainerService, Service} from './service'
 export {Logger} from './logger'
 export {Store} from './store'
+
+export * as Extension from './extension'
 
 /**
  * Compilation mode
@@ -71,18 +72,3 @@ import {Plugin} from './services/extensions'
 export interface Plugins extends Partial<Record<string, Plugin>> {}
 export {Plugin}
 
-/**
- * Factory
- * @public
- */
-export interface Factory<P extends any[], T> {
-  (...args: P): T
-}
-
-/**
- * Async factory
- * @public
- */
-export interface AsyncFactory<P extends any[], T> {
-  (...args: P): Promise<T>
-}
