+++ sources/@roots/bud-framework/src/services/extensions/index.ts
@@ -1,2 +1,2 @@
-export {Module, Plugin} from './extension'
+export {Module, Plugin} from '../../extension'
 export {Service} from './extensions.service'
