+++ sources/@roots/bud-hooks/src/index.ts
@@ -1,7 +1,7 @@
 /**
  * Hooks system used for framework eventing.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @remarks
