+++ sources/@roots/bud-imagemin/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Add image optimization support to Bud projects
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
