+++ sources/@roots/bud-library/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Adds dynamic link library (DLL) support to Bud
 
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
