+++ sources/@roots/bud-postcss/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Adds PostCSS support to Bud
 
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
