+++ sources/@roots/bud-preset-recommend/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Recommended preset configuration for Bud.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
