+++ sources/@roots/bud-prettier/src/index.ts
@@ -10,7 +10,7 @@
  *
  * It will be developed further as the Framework matures.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
