+++ sources/@roots/bud-purgecss/src/index.ts
@@ -13,7 +13,7 @@
  * })
  * ```
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
