+++ sources/@roots/bud-react/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Adds support for react to bud projects.
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  *
