+++ sources/@roots/bud-sass/src/index.ts
@@ -1,7 +1,7 @@
 /**
  * Add sass support to Bud projects
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
