+++ sources/@roots/bud-server/src/index.ts
@@ -1,7 +1,7 @@
 /**
  *  Development server
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
