+++ sources/@roots/bud-stylelint/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Add stylelint support to Bud
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
