+++ sources/@roots/bud-tailwindcss/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Adds tailwindcss support to Bud
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
  *
  * @packageDocumentation
