+++ sources/@roots/bud-vue/src/index.ts
@@ -4,7 +4,7 @@
 /**
  * Adds vue sfc support to Bud projects
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
 
  * @packageDocumentation
