+++ sources/@roots/bud-wordpress-externals/src/index.ts
@@ -7,7 +7,7 @@
  * This package replaces source code references to WordPress provided packages and collects package
  * references in the {@link @roots/bud-entrypoints# | entrypoints manifest}
  *
- * @see https://roots.io/bud
+ * @see https://bud.js.org
  * @see https://github.com/roots/bud
 
  * @remarks
